Explain workaround in Rake task and README

+++ Rakefile
@@ -1,5 +1,8 @@
 desc "Publish"
 task :publish do
+  unless ENV["RUBYOPT"].to_s.include?("logger")
+    abort "Oh, I could not see a workaround RUBYOPT=-rlogger on the command line. See the README if stuff does not work."
+  end
   # Delete the old build
   FileUtils.rm_r('build') if File.exist?('build')
